+++ GW2EIEvtcParser/EIData/CombatReplay/Decorations/MovingPlatformDecoration.cs
@@ -8,8 +8,8 @@ namespace GW2EIEvtcParser.EIData
         public int Width { get; }
         public int Height { get; }
 
-        public List<(double x, double y, double z, double angle, double opacity, int time)> Positions { get; } =
-            new List<(double x, double y, double z, double angle, double opacity, int time)>();
+        public List<(float x, float y, float z, float angle, float opacity, int time)> Positions { get; } =
+            new List<(float x, float y, float z, float angle, float opacity, int time)>();
 
         public MovingPlatformDecoration(string image, int width, int height, (int start, int end) lifespan) : base(lifespan)
         {
@@ -18,9 +18,9 @@ namespace GW2EIEvtcParser.EIData
             Height = height;
         }
 
-        public void AddPosition(double x, double y, double z, double angle, double opacity, int time)
+        public void AddPosition(float x, float y, float z, double angle, double opacity, int time)
         {
-            Positions.Add((x, y, z, angle, opacity, time));
+            Positions.Add((x, y, z, (float)angle, (float)opacity, time));
         }
 
         public override GenericDecorationCombatReplayDescription GetCombatReplayDescription(CombatReplayMap map, ParsedEvtcLog log)
